// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/primaryGeographicalAreaAC.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BaseModelAC } from './baseModelAC';
import { LocationAC } from './locationAC';
import { RegionAC } from './regionAC';

export interface PrimaryGeographicalAreaAC extends BaseModelAC { 
    entityId: string;
    auditableEntity?: string;
    regionId: string;
    regionString?: string;
    entityCountryId: string;
    countryString?: string;
    entityStateId: string;
    stateString?: string;
    locationId: string;
    locationString?: string;
    regionACList?: Array<RegionAC>;
    locationACList?: Array<LocationAC>;
}